WIP: Refactor of IR impl
This is a rough refactor of the code from Lukas removing some unnecessary code and trying to get it in the (imo) right direction.
I am still *really* not happy with the shared/weak_ptr stuff since the dependencies are rather unclear to see through and i'd like a more easily understandable "memory graph" there but that's for tomorrow to think about. We definitely need to think about wether we want to just make two big vectors for functions/basic-blocks and then just index into them and use that as ids instead of dragging around pointers everywhere.
I also noticed that the "predecessor mapping" doesn't really seem to make a lot of sense. It should probably just be replaced with a more straight-forward input-list for the basic block.
I also replaced the operation-list in the basic-block with a list of the variables of the basic-block (see the TODO's in basic_block.h), also needs discussion.
The Control-Flow ops are untouched for now. They probably need changing as well.
Oh, it also prints a bit now.