refactor to a more suitable solution
refactor of the code to a state where dependencies, memory management and usability are (imo) clearer.
see the example code in main.cpp.
also added some more printing support.
the list of variables instead of list of ops (or both) issue is still present in the basic blocks; still needs to be looked at if it's okay this way