fix broken ref counting
when cfops where copied around (for example on a vector resize) the ref counting could be done improperly
this commit fixes this and also adds some safeguards to make breaking the ref count harder, e.g. for optimizing passes in the ir which now have to make extra calls to get a modifiable version of the vectors which should help keep ref counting in mind.