Rewrite leaky CfOp::target_inputs()
CfOp::target_inputs() sometimes copied RefPtrs onto a static vector in
the method and returned a reference to that vector. This meant that the
RefPtrs were potentially left in that vector for the entire runtime,
therefore resulting in a leak.

Since no caller actually needed the RefPtr itself, only copying the
SSAVar* is enough, and prevents the leak.